fixed issue #760
